:truck: Moved forms,views,urls to specific models
