+++ Qatar_World_Cup_2022/AppQatarWC2022/album/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+
+from AppQatarWC2022.album.views import my_album
+
+urlpatterns = [
+    path("my_album", my_album, name="my_album"),
+]
+
+
