+++ Qatar_World_Cup_2022/AppQatarWC2022/album/views/__init__.py
@@ -1,0 +1,1 @@
+from .my_album_view import my_album
