+++ Qatar_World_Cup_2022/AppQatarWC2022/album/views/my_album_view.py
@@ -1,0 +1,13 @@
+from django.shortcuts import render
+from AppQatarWC2022.models import app
+from django.contrib.auth.decorators import login_required
+
+working_context = {'working_context':app.working_context()}
+
+@login_required
+def my_album(request):
+    if request.method =="POST":
+        pass
+    else:
+        app.working_context().refresh_album()
+        return render(request,"my_album.html",working_context)
