+++ Qatar_World_Cup_2022/AppQatarWC2022/countries/forms/__init__.py
@@ -1,0 +1,1 @@
+from .country_form import CountryRegistration
