+++ Qatar_World_Cup_2022/AppQatarWC2022/generic_modals/views/__init__.py
@@ -1,0 +1,1 @@
+from .generic_modals_view import element_registration,element_unregistration,element_update
