+++ Qatar_World_Cup_2022/AppQatarWC2022/promo_codes/forms/__init__.py
@@ -1,0 +1,1 @@
+from .promo_code_form import PromoCodeRegistration
