+++ Qatar_World_Cup_2022/AppQatarWC2022/promo_codes/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from AppQatarWC2022.promo_codes.views import promo_codes
+
+urlpatterns = [
+    path('promo_codes/',promo_codes,name='promo_codes'),
+]
