+++ Qatar_World_Cup_2022/AppQatarWC2022/promo_codes/views/__init__.py
@@ -1,0 +1,1 @@
+from .promo_codes_view import promo_codes
