+++ Qatar_World_Cup_2022/AppQatarWC2022/stickers/forms/__init__.py
@@ -1,0 +1,1 @@
+from .player_sticker_form import PlayerStickerRegistration
