+++ Qatar_World_Cup_2022/AppQatarWC2022/stickers/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+
+from AppQatarWC2022.stickers.views import player_stickers
+
+urlpatterns = [
+    path('player_stickers/',player_stickers,name='player_stickers'),
+]
