+++ Qatar_World_Cup_2022/AppQatarWC2022/stickers/views/__init__.py
@@ -1,0 +1,1 @@
+from .player_stickers_view import player_stickers
