+++ Qatar_World_Cup_2022/AppQatarWC2022/users/forms/__init__.py
@@ -1,0 +1,2 @@
+from .sign_in_form import SignIn
+from .user_form import UserRegistration
