+++ Qatar_World_Cup_2022/AppQatarWC2022/users/forms/sign_in_form.py
@@ -1,0 +1,6 @@
+from django import forms
+from django.contrib.auth.forms import AuthenticationForm
+
+class SignIn(AuthenticationForm):
+    username = forms.CharField(label='Usuario',widget=forms.TextInput(attrs={'class':'form-control'}))
+    password= forms.CharField(label='Contraseña',widget=forms.PasswordInput(attrs={'class':'form-control'}))
