+++ Qatar_World_Cup_2022/AppQatarWC2022/users/views/__init__.py
@@ -1,0 +1,2 @@
+from .login_view import login_request
+from .sign_up_view import sign_up
